@@ -17,48 +17,47 @@
 		"test:unit": "vitest"
 	},
 	"devDependencies": {
-		"@axe-core/playwright": "^4.9.1",
-		"@floating-ui/dom": "^1.6.10",
-		"@iconify-json/bi": "^1.1.24",
-		"@playwright/test": "^1.46.1",
-		"@sveltejs/adapter-node": "^5.2.2",
-		"@sveltejs/kit": "^2.5.24",
-		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@axe-core/playwright": "^4.10.1",
+		"@eslint/compat": "^1.2.9",
+		"@floating-ui/dom": "^1.6.13",
+		"@iconify-json/bi": "^1.2.2",
+		"@playwright/test": "^1.51.0",
+		"@sveltejs/adapter-node": "^5.2.12",
+		"@sveltejs/kit": "^2.20.8",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@tailwindcss/vite": "^4.1.3",
 		"@types/eslint": "^9.6.0",
 		"@types/jmespath": "^0.15.2",
 		"@types/js-cookie": "^3.0.6",
-		"autoprefixer": "^10.4.20",
 		"codemirror-lang-lxlquery": "^0.1.0",
 		"cssnano": "^7.0.5",
-		"eslint": "^9.9.0",
-		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-svelte": "^2.43.0",
+		"eslint": "^9.22.0",
+		"eslint-config-prettier": "^10.1.1",
+		"eslint-plugin-svelte": "^3.5.1",
 		"estree-walker": "^3.0.3",
-		"globals": "^15.9.0",
+		"globals": "^16.0.0",
 		"jmespath": "^0.16.0",
 		"js-cookie": "^3.0.5",
 		"lxljs": "^1.1.0",
 		"magic-string": "^0.30.11",
 		"mdsvex": "^0.12.3",
 		"nprogress": "^0.2.0",
-		"postcss": "^8.4.41",
-		"postcss-import": "^16.1.0",
-		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.6",
-		"prettier-plugin-tailwindcss": "^0.6.6",
+		"prettier": "^3.5.3",
+		"prettier-plugin-svelte": "^3.3.3",
+		"prettier-plugin-tailwindcss": "^0.6.11",
 		"supersearch": "^0.0.1",
-		"svelte": "^5.18.0",
-		"svelte-check": "^4.0.4",
-		"tailwindcss": "^3.4.10",
+		"svelte": "^5.28.2",
+		"svelte-check": "^4.1.7",
+		"tailwindcss": "^4.1.3",
 		"tslib": "^2.6.3",
-		"typescript": "^5.5.4",
-		"typescript-eslint": "^8.1.0",
-		"unplugin-icons": "^0.19.2",
-		"vanilla-cookieconsent": "^3.0.1",
-		"vite": "^5.4.6",
-		"vitest": "^2.0.5"
+		"typescript": "^5.8.2",
+		"typescript-eslint": "^8.26.1",
+		"unplugin-icons": "^22.1.0",
+		"vanilla-cookieconsent": "^3.1.0",
+		"vite": "^6.2.7",
+		"vitest": "^3.0.8"
 	},
 	"dependencies": {
-		"dotenv": "^16.4.5"
+		"dotenv": "^16.4.7"
 	}
 }
