@@ -19,22 +19,22 @@
 	"types": "dist/index.d.ts",
 	"sideEffects": false,
 	"dependencies": {
-		"@codemirror/language": "^6.0.0",
+		"@codemirror/language": "^6.11.0",
 		"@lezer/highlight": "^1.0.0",
 		"@lezer/lr": "^1.0.0"
 	},
 	"devDependencies": {
-		"@lezer/generator": "^1.0.0",
-		"eslint": "^9.14.0",
-		"eslint-config-prettier": "^9.1.0",
-		"globals": "^15.12.0",
-		"prettier": "^3.3.3",
-		"rollup": "^2.60.2",
+		"@lezer/generator": "^1.7.2",
+		"eslint": "^9.22.0",
+		"eslint-config-prettier": "^10.1.1",
+		"globals": "^16.0.0",
+		"prettier": "^3.5.3",
+		"rollup": "^2.79.2",
 		"rollup-plugin-dts": "^4.0.1",
 		"rollup-plugin-ts": "^3.0.2",
-		"typescript": "^4.3.4",
-		"typescript-eslint": "^8.13.0",
-		"vitest": "^2.1.4"
+		"typescript": "^4.9.5",
+		"typescript-eslint": "^8.26.1",
+		"vitest": "^3.0.8"
 	},
 	"license": "MIT"
 }
