@@ -38,31 +38,32 @@
 		"svelte": "^5.0.0"
 	},
 	"devDependencies": {
-		"@codemirror/commands": "^6.7.1",
-		"@codemirror/state": "^6.4.1",
-		"@codemirror/view": "^6.34.1",
-		"@playwright/test": "^1.45.3",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/package": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@codemirror/commands": "^6.8.0",
+		"@codemirror/state": "^6.5.2",
+		"@codemirror/view": "^6.36.4",
+		"@eslint/compat": "^1.2.9",
+		"@playwright/test": "^1.51.0",
+		"@sveltejs/adapter-auto": "^6.0.0",
+		"@sveltejs/kit": "^2.20.8",
+		"@sveltejs/package": "^2.3.11",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"@types/eslint": "^9.6.0",
 		"codemirror": "^6.0.1",
 		"codemirror-lang-lxlquery": "^0.1.0",
-		"eslint": "^9.7.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.36.0",
+		"eslint": "^9.22.0",
+		"eslint-config-prettier": "^10.1.1",
+		"eslint-plugin-svelte": "^3.5.1",
 		"esm-env": "^1.2.1",
-		"globals": "^15.0.0",
-		"prettier": "^3.3.2",
+		"globals": "^16.0.0",
+		"prettier": "^3.5.3",
 		"prettier-plugin-css-order": "^2.1.2",
-		"prettier-plugin-svelte": "^3.2.6",
+		"prettier-plugin-svelte": "^3.3.3",
 		"publint": "^0.2.0",
-		"svelte": "^5.18.0",
-		"svelte-check": "^4.0.0",
-		"typescript": "^5.0.0",
-		"typescript-eslint": "^8.0.0",
-		"vite": "^5.0.11",
-		"vitest": "^2.0.4"
+		"svelte": "^5.28.2",
+		"svelte-check": "^4.1.7",
+		"typescript": "^5.8.2",
+		"typescript-eslint": "^8.26.1",
+		"vite": "^6.2.7",
+		"vitest": "^3.0.8"
 	}
 }
