@@ -1,31 +1,36 @@
 <script lang="ts">
 	import jmespath from 'jmespath';
+	import { page } from '$app/state';
 	import { replaceState } from '$app/navigation';
 
 	import type { ResourceData } from '$lib/types/resourceData';
+	import { getUserSettings } from '$lib/contexts/userSettings';
 	import { getResourceId } from '$lib/utils/resourceData';
 	import { relativizeUrl } from '$lib/utils/http';
-	import { getHoldingsLink, handleClickHoldings } from '$lib/utils/holdings';
+	import { getHoldingsLink, getMyLibsFromHoldings, handleClickHoldings } from '$lib/utils/holdings';
 
-	import BiChevronRight from '~icons/bi/chevron-right';
-	import DecoratedData from '$lib/components/DecoratedData.svelte';
 	import InstancesListContent from './InstancesListContent.svelte';
-	import { page } from '$app/stores';
+	import DecoratedData from '$lib/components/DecoratedData.svelte';
+	import MyLibrariesIndicator from '$lib/components/MyLibsHoldingIndicator.svelte';
+	import BiChevronRight from '~icons/bi/chevron-right';
 
 	/**
 	 * TODO:
 	 * - [] Replace jmespath (used for queriyng data for the columns) with something more home-baked (with smaller bundle-size). Another alternative could be querying and preparing the column data server-side?
 	 * - [] Add tests (e.g. rows should be expandable and permalink should work, and that opened state should be saved when navigating forward and backwards). The tests should probably be placed inside +page.svelte...
 	 */
+	type InstancesListProps = {
+		data: ResourceData;
+		columns: { header: string; data: string }[];
+	};
+	const { data, columns }: InstancesListProps = $props();
 
-	let instancesList: HTMLUListElement;
-
-	export let data: ResourceData;
-	export let columns: { header: string; data: string }[];
+	let instancesList: HTMLUListElement | undefined = $state();
+	let userSettings = getUserSettings();
 
 	function handleToggleDetails(state: { expandedInstances?: string[] }) {
 		let openIds: string[] = [];
-		instancesList.querySelectorAll('details[open]').forEach((openElement) => {
+		instancesList?.querySelectorAll('details[open]').forEach((openElement) => {
 			const parentId = openElement.parentElement?.id;
 			if (parentId) {
 				openIds = [...openIds, parentId];
@@ -69,24 +74,24 @@
 <div>
 	{#if Array.isArray(data) && data.length > 1}
 		<div class="flex items-center justify-between pb-4">
-			<h2 class="capitalize text-4-cond-bold">{$page.data.t('search.editions')}</h2>
+			<h2 class="font-heading text-2xl capitalize">{page.data.t('search.editions')}</h2>
 			<a
-				href={getCollapseAllUrl($page.url)}
+				href={getCollapseAllUrl(page.url)}
 				data-sveltekit-preload-data="false"
-				class="close-all text-disabled:text-disabled text-xs sm:text-sm"
-				on:click={(event) => {
+				class="close-all link-subtle text-xs sm:text-sm"
+				onclick={(event) => {
 					event.preventDefault();
-					replaceState(getCollapseAllUrl($page.url), { ...$page.state, expandedInstances: [] });
+					replaceState(getCollapseAllUrl(page.url), { ...page.state, expandedInstances: [] });
 				}}
 			>
-				{$page.data.t('general.collapseAll')}
+				{page.data.t('general.collapseAll')}
 			</a>
 		</div>
 	{/if}
 
 	{#if Array.isArray(data) && data.length > 1}
-		<div class="column-headers mb-2 grid gap-2 text-sm font-bold">
-			{#each columns as { header: columnHeader }}
+		<div class="column-headers mb-2 grid gap-2 text-sm font-medium">
+			{#each columns as { header: columnHeader }, index (index)}
 				<div class="flex flex-1 first:pl-0">
 					{columnHeader}
 				</div>
@@ -95,41 +100,48 @@
 		<ul bind:this={instancesList}>
 			{#each data as item (item['@id'])}
 				{@const id = relativizeUrl(getResourceId(item))}
-				<li {id} class="border-t border-t-primary/16">
+				<li {id} class="border-neutral border-t">
 					<details
-						open={$page.state.expandedInstances?.includes(id) ||
-							$page.url.searchParams.getAll('expanded').includes(id) ||
+						open={page.state.expandedInstances?.includes(id) ||
+							page.url.searchParams.getAll('expanded').includes(id) ||
 							data.length === 1}
-						on:toggle={() => handleToggleDetails($page.state)}
+						ontoggle={() => handleToggleDetails(page.state)}
 					>
 						<summary
-							class="grid min-h-11 items-center gap-2 align-middle text-sm hover:bg-pill/16 md:text-base"
-							on:keydown={handleSummaryKeydown}
+							class="hover:bg-primary-50 grid min-h-11 items-center gap-2 align-middle text-sm"
+							onkeydown={handleSummaryKeydown}
 						>
-							<span class="arrow w-4">
+							<span class="arrow text-subtle w-4 origin-center rotate-0 transition-transform">
 								<BiChevronRight />
 							</span>
-							{#each columns as { data: columnData }}
+							{#each columns as { data: columnData }, index (index)}
 								<div class="flex flex-1 items-center">
 									<DecoratedData data={jmespath.search(item, columnData)} />
 								</div>
 							{/each}
 							<div class="text flex flex-1 items-center justify-end text-sm">
-								{#if id && $page.data.holdingsByInstanceId[id]}
+								{#if id && page.data.holdingsByInstanceId[id]}
+									{@const myLibsWithHolding = getMyLibsFromHoldings(
+										userSettings.myLibraries,
+										page.data.holdingsByInstanceId[id]
+									)}
+									{#if myLibsWithHolding.length}
+										<MyLibrariesIndicator libraries={myLibsWithHolding} />
+									{/if}
 									<a
-										href={getHoldingsLink($page.url, id)}
-										class="flex items-center self-center text-xs sm:text-sm"
+										href={getHoldingsLink(page.url, id)}
+										class="link-subtle flex items-center self-center text-xs sm:text-sm"
 										data-sveltekit-preload-data="false"
-										on:click={(event) => handleClickHoldings(event, $page.state, id)}
+										onclick={(event) => handleClickHoldings(event, page.state, id)}
 									>
-										{$page.data.holdingsByInstanceId[id].length}
-										{$page.data.holdingsByInstanceId[id].length === 1
-											? $page.data.t('holdings.library')
-											: $page.data.t('holdings.libraries')}
+										{page.data.holdingsByInstanceId[id].length}
+										{page.data.holdingsByInstanceId[id].length === 1
+											? page.data.t('holdings.library')
+											: page.data.t('holdings.libraries')}
 									</a>
 								{:else}
 									<span>
-										{$page.data.t('holdings.availableAt')} 0 {$page.data.t('holdings.libraries')}
+										{page.data.t('holdings.availableAt')} 0 {page.data.t('holdings.libraries')}
 									</span>
 								{/if}
 							</div>
@@ -140,13 +152,13 @@
 			{/each}
 		</ul>
 	{:else}
-		<div class="max-w-content">
-			<InstancesListContent data={data[0]} id={relativizeUrl(getResourceId(data[0]))} />
-		</div>
+		<InstancesListContent data={data[0]} id={relativizeUrl(getResourceId(data[0]))} />
 	{/if}
 </div>
 
 <style lang="postcss">
+	@reference "../../../../app.css";
+
 	.column-headers,
 	summary {
 		@apply grid-cols-instance-list gap-2;
@@ -160,16 +172,7 @@
 		grid-column: span 2;
 	}
 
-	details[open] > summary {
-		@apply bg-pill/8 hover:bg-pill/16;
-	}
-
 	details[open] .arrow {
 		@apply rotate-90;
-	}
-
-	.arrow {
-		transform-origin: center;
-		@apply rotate-0 transition-transform;
 	}
 </style>
