@@ -1,10 +1,14 @@
 <script lang="ts">
 	import '../app.css';
+	import { afterNavigate, beforeNavigate } from '$app/navigation';
+	import { page } from '$app/state';
+	import { setUserSettings } from '$lib/contexts/userSettings';
 	import NProgress from 'nprogress';
 	import '$lib/styles/nprogress.css';
-	import { navigating } from '$app/stores';
 	import Matomo from '$lib/components/Matomo.svelte';
 	import CookieConsent from '$lib/components/CookieConsent.svelte';
+
+	const { children } = $props();
 
 	NProgress.configure({
 		//https://github.com/rstacruz/nprogress#configuration
@@ -12,29 +16,30 @@
 		showSpinner: false
 	});
 
-	let progressBarTimeout: number | undefined = undefined;
-	const progressDelay = 200;
+	setUserSettings(page.data.userSettings);
 
-	$: {
-		if ($navigating) {
-			clearTimeout(progressBarTimeout);
-			progressBarTimeout = setTimeout(NProgress.start, progressDelay);
-		}
-		if (!$navigating) {
-			clearTimeout(progressBarTimeout);
-			NProgress.done();
-		}
-	}
+	let progressBarTimeout: ReturnType<typeof setTimeout> | undefined = undefined;
+	const progressDelay = 300;
+
+	beforeNavigate(() => {
+		clearTimeout(progressBarTimeout);
+		progressBarTimeout = setTimeout(NProgress.start, progressDelay);
+	});
+
+	afterNavigate(() => {
+		clearTimeout(progressBarTimeout);
+		NProgress.done();
+	});
 </script>
 
 <div class="body-layout grid min-h-screen">
 	<Matomo>
-		<slot />
+		{@render children()}
 		<CookieConsent />
 	</Matomo>
 </div>
 
-<style lang="postcss">
+<style>
 	.body-layout {
 		grid-template-rows: min-content 1fr min-content;
 	}
