@@ -14,7 +14,12 @@ const config: PlaywrightTestConfig = {
 	],
 	webServer: {
 		command: 'npm run build && npm run preview',
-		port: 4173
+		port: 4173,
+		reuseExistingServer: !process.env.CI
+	},
+	retries: 1,
+	use: {
+		trace: process.env.CI ? 'on-first-retry' : 'off'
 	},
 	testDir: 'tests',
 	testMatch: /(.+\.)?(test|spec)\.[jt]s/
