@@ -2,8 +2,9 @@
 	import { onMount } from 'svelte';
 	import * as CookieConsent from 'vanilla-cookieconsent';
 	import 'vanilla-cookieconsent/dist/cookieconsent.css';
-	import svTranslations from '$lib/i18n/locales/cookieConsent.sv';
-	import enTranslations from '$lib/i18n/locales/cookieConsent.en';
+	import '$lib/styles/cookieconsent.css';
+	import svTranslations from '$lib/i18n/cookieConsent/sv';
+	import enTranslations from '$lib/i18n/cookieConsent/en';
 	import { page } from '$app/stores';
 	import { getMatomoTracker } from '$lib/contexts/matomo';
 
@@ -59,25 +60,3 @@
 		CookieConsent.run(config);
 	});
 </script>
-
-<style lang="postcss">
-	/* heading */
-	:global(#cc-main .cm__title, #cc-main .pm__title) {
-		@apply text-primary text-4-cond-bold;
-	}
-
-	/* subheading */
-	:global(#cc-main .pm__section-title) {
-		@apply text-primary text-3-cond;
-	}
-
-	/* body text */
-	:global(#cc-main .cm__desc, #cc-main .pm__section-desc) {
-		@apply text-primary text-2-regular;
-	}
-
-	/* btns */
-	:global(#cc-main .cm__btn, #cc-main .pm__btn) {
-		@apply text-3-cond;
-	}
-</style>
