@@ -1,9 +1,7 @@
 <script lang="ts">
-	import { ShowLabelsOptions } from '$lib/types/decoratedData';
 	import type { DisplayMapping, SearchOperators } from '$lib/types/search';
 
 	import { page } from '$app/stores';
-	import DecoratedData from '$lib/components/DecoratedData.svelte';
 	import { getModalContext } from '$lib/contexts/modal';
 	import BiXLg from '~icons/bi/x-lg';
 	import BiPencil from '~icons/bi/pencil';
@@ -49,9 +47,9 @@
 </script>
 
 <ul class="flex flex-wrap items-center gap-2">
-	{#each mapping as m}
+	{#each mapping as m, index (`${m['@id']}-${index}`)}
 		<li
-			class="mapping-item {m.children ? 'pill-group' : 'pill'} pill-{m.operator}"
+			class="mapping-item {m.children ? 'pill-group' : 'btn btn-accent'} pill-{m.operator}"
 			class:wildcard={m.operator === 'equals' && m.display === '*'}
 			class:outer={depth === 0}
 			class:free-text={m?.['@id'] === 'https://id.kb.se/vocab/textQuery'}
@@ -61,13 +59,13 @@
 			{:else if m.operator === 'existence' || m.operator === 'notExistence'}
 				{@const symbol = getRelationSymbol(m.operator)}
 				<span class="pill-relation">{symbol}</span>
-				<div class="pill-label inline text-2-regular">{m.label}</div>
+				<div class="pill-label inline">{m.label}</div>
 			{:else if 'label' in m && 'display' in m}
 				{@const symbol = getRelationSymbol(m.operator)}
-				<div class="pill-label inline text-2-regular">{m.label}</div>
+				<div class="pill-label inline">{m.label}</div>
 				<span class="pill-relation">{symbol}</span>
 				<span class="pill-value">
-					<DecoratedData data={m.display} showLabels={ShowLabelsOptions['Never']} />
+					{m.displayStr}
 				</span>
 			{/if}
 			{#if 'up' in m && (!m.children || depth > 0)}
@@ -83,11 +81,11 @@
 			{/if}
 		</li>
 		{#if parentOperator}
-			<li class="pill-between pill-between-{parentOperator}">{parentOperator}</li>
+			<li class="pill-between pill-between-{parentOperator} text-xs">{parentOperator}</li>
 		{/if}
 		{#if 'up' in m && m.children && depth === 0}
 			<li class="pill-remove">
-				<a class="button-ghost button-negative" href={m.up?.['@id']}>
+				<a href={m.up?.['@id']} class="btn btn-primary">
 					<BiTrash aria-hidden="true" />
 					{$page.data.t('search.clearFilters')}
 				</a>
@@ -96,12 +94,7 @@
 	{/each}
 	{#if !inModal && showEditButton && depth === 0}
 		<li>
-			<a
-				class="button-ghost"
-				class:active={editActive}
-				data-sveltekit-replacestate
-				href={toggleEditUrl}
-			>
+			<a class:active={editActive} data-sveltekit-replacestate href={toggleEditUrl}>
 				{#if editActive}
 					<BiPencilFill aria-hidden="true" />
 				{:else}
@@ -114,37 +107,27 @@
 </ul>
 
 <style lang="postcss">
+	@reference "../../../app.css";
+
 	.mapping-item {
-		@apply rounded-md px-4 py-2 brightness-100 text-3-cond-bold;
-		transition: filter 0.1s ease;
+		/* @apply rounded-md px-4 py-2 brightness-100;
+		transition: filter 0.1s ease; */
 	}
 
 	.mapping-item:has(> .pill-remove:hover) {
-		@apply brightness-[.85];
+		/* @apply brightness-[.85]; */
 	}
 
 	.pill {
-		@apply bg-positive-inv text-primary-inv;
-
 		& .pill-label,
 		.pill-relation {
-			@apply text-secondary-inv;
-		}
-	}
-
-	.pill-equals {
-		:global(.text-secondary) {
-			@apply text-secondary-inv;
 		}
 	}
 
 	.pill-notEquals,
 	.pill-notExistence {
-		@apply bg-negative text-primary;
-
 		& .pill-label,
 		.pill-relation {
-			@apply text-secondary;
 		}
 	}
 
@@ -153,7 +136,7 @@
 	}
 
 	.pill-group {
-		@apply flex items-center gap-2 bg-pill/8 p-0 pr-4;
+		@apply flex items-center gap-2 p-0 pr-4;
 
 		&.outer {
 			@apply bg-transparent;
@@ -162,7 +145,7 @@
 
 	.pill-between,
 	.pill-relation {
-		@apply uppercase text-primary text-2-regular;
+		@apply uppercase;
 	}
 
 	.pill-between-and,
