@@ -9,11 +9,12 @@
 	import SearchRelated from './SearchRelated.svelte';
 	import IconSliders from '~icons/bi/sliders';
 	import BiChevronDown from '~icons/bi/chevron-down';
+	import BiSortDown from '~icons/bi/sort-down';
 	import type { SearchResult, DisplayMapping } from '$lib/types/search';
-	import { shouldShowMapping } from '$lib/utils/search';
 
 	let showFiltersModal = false;
 	export let searchResult: SearchResult;
+	export let showMapping: boolean = false;
 
 	$: sortOrder = $page.url.searchParams.get('_sort');
 	const sortOptions = [
@@ -51,22 +52,24 @@
 
 <slot />
 {#if searchResult}
-	{@const facets = searchResult.facetGroups}
+	{@const facets = searchResult.facetGroups || []}
+	{@const predicates = searchResult.predicates || []}
 	{@const numHits = searchResult.totalItems}
 	{@const filterCount = getFiltersCount(searchResult.mapping)}
-	{#if searchResult.predicates.length}
+	{#if predicates.length}
 		<nav
-			class="border-b border-primary/16 px-4 md:flex lg:px-6"
+			class="border-neutral border-b px-4 lg:flex 2xl:px-6"
 			aria-label={$page.data.t('search.selectedFilters')}
 		>
-			<ul class="flex flex-wrap items-center gap-2">
-				<li class="tab-header max-w-80 truncate font-bold">{$page.data.title}</li>
-				<span class="tab-header">{$page.data.t('search.occursAs')}</span>
-
-				{#each searchResult.predicates as p}
+			<ul class="flex flex-wrap items-center gap-2 text-sm">
+				<li class="block max-w-80 truncate py-4 whitespace-nowrap">
+					<span class="font-medium">{$page.data.title}</span>
+				</li>
+				<li>{$page.data.t('search.occursAs')}</li>
+				{#each predicates as p}
 					<li>
 						<a
-							class="tab"
+							class="flex flex-nowrap items-center gap-1 py-4 pr-3.5 pl-4 lowercase no-underline"
 							class:active={true}
 							class:tab-selected={p.selected}
 							data-sveltekit-replacestate
@@ -74,7 +77,7 @@
 						>
 							{p.str}
 							<span
-								class="mb-px rounded-sm bg-pill/4 px-1 text-sm text-secondary md:text-xs lg:text-sm"
+								class="badge badge-accent"
 								aria-label="{p.totalItems} {$page.data.t('search.hits')}">{p.totalItems}</span
 							>
 						</a>
@@ -83,15 +86,15 @@
 			</ul>
 		</nav>
 	{/if}
-	{#if shouldShowMapping(searchResult.mapping)}
+	{#if showMapping}
 		<nav
-			class="hidden md:flex md:px-6 md:pb-0 md:pt-4"
+			class="hidden lg:flex lg:px-6 lg:pt-4 lg:pb-0"
 			aria-label={$page.data.t('search.selectedFilters')}
 		>
 			<SearchMapping mapping={searchResult.mapping} />
 		</nav>
 	{/if}
-	<div class="relative gap-y-4 find-layout md:page-padding">
+	<div class="find-layout relative gap-y-4 p-4 sm:px-6">
 		{#if showFiltersModal}
 			<Modal position="left" close={toggleFiltersModal}>
 				<span slot="title">
@@ -101,36 +104,34 @@
 				<Filters {facets} mapping={searchResult.mapping} />
 			</Modal>
 		{/if}
-		<div class="filters hidden md:block" id="filters">
+		<div class="filters hidden lg:block" id="filters">
 			<Filters {facets} mapping={searchResult.mapping} />
 		</div>
 
-		<div class="results max-w-content">
+		<div class="results">
 			<div
-				class="toolbar flex min-h-14 items-center justify-between page-padding md:min-h-fit md:p-0 md:pb-4"
+				class="toolbar flex items-center justify-between pb-2"
 				class:has-search={$page.params.fnurgel}
 			>
 				<a
 					href={`${$page.url.pathname}?${$page.url.searchParams.toString()}#filters`}
-					class="filter-modal-toggle button-ghost md:hidden"
+					class="filter-modal-toggle btn btn-primary lg:hidden"
 					aria-label={$page.data.t('search.filters')}
 					on:click|preventDefault={toggleFiltersModal}
 				>
-					<IconSliders width={20} height={20} />
+					<IconSliders class="text-base" />
 					{$page.data.t('search.filters')}
 					{#if filterCount}
-						<span
-							class="flex h-5 w-5 items-center justify-center rounded-full bg-pill text-xs font-bold leading-none text-primary-inv"
-						>
+						<span class="badge badge-accent">
 							{filterCount}
 						</span>
 					{/if}
 				</a>
-				<span class="hits pt-4 text-secondary md:pt-0" role="status" data-testid="result-info">
+				<span class="hits text-2xs pt-4 lg:pt-0" role="status" data-testid="result-info">
 					{#if numHits && numHits > 0}
 						<span class="hits-count">
 							{#if numHits > searchResult.itemsPerPage}
-								<span class="text-3-cond-bold">
+								<span class="font-medium">
 									{(searchResult.itemOffset + 1).toLocaleString($page.data.locale)}
 									-
 									{Math.min(
@@ -140,7 +141,7 @@
 								</span>
 								{$page.data.t('search.hitsOf')}
 							{/if}
-							<span class="text-3-cond-bold">
+							<span class="font-medium">
 								{numHits.toLocaleString($page.data.locale)}
 							</span>
 							{#if $page.data.instances}
@@ -155,7 +156,10 @@
 						<span class="suggest">
 							{#each searchResult._spell as suggestion (suggestion.label)}
 								{$page.data.t('search.didYouMean')}
-								<a href={suggestion.view['@id'].replace('_spell=true', '_spell=false')}>
+								<a
+									href={suggestion.view['@id'].replace('_spell=true', '_spell=false')}
+									class="link-subtle"
+								>
 									<!-- eslint-disable-next-line svelte/no-at-html-tags -->
 									{@html suggestion.labelHtml}</a
 								>?
@@ -165,7 +169,7 @@
 				</span>
 				<div class="search-related flex justify-start">
 					{#if $page.params.fnurgel}
-						{@const activePredicate = searchResult.predicates.filter((p) => p.selected)}
+						{@const activePredicate = predicates.filter((p) => p.selected)}
 						<SearchRelated view={activePredicate[0].view} />
 					{/if}
 				</div>
@@ -174,25 +178,33 @@
 						class="sort-select flex flex-col items-end justify-self-end"
 						data-testid="sort-select"
 					>
-						<label class="pr-6 text-secondary text-2-regular" for="search-sort">
+						<label class="sr-only" for="search-sort">
 							{$page.data.t('sort.sort')}
 						</label>
 						<div class="relative">
-							<select id="search-sort" form="main-search" on:change={handleSortChange}>
+							<span class="text-subtle pointer-events-none absolute top-0 p-2">
+								<BiSortDown aria-hidden="true" />
+							</span>
+							<select
+								id="search-sort"
+								class="btn btn-primary"
+								form="main-search"
+								on:change={handleSortChange}
+							>
 								{#each sortOptions as option}
 									<option value={option.value} selected={option.value === sortOrder}
 										>{option.label}</option
 									>
 								{/each}
 							</select>
-							<span class="pointer-events-none absolute right-0 top-[5px]">
-								<BiChevronDown aria-hidden="true" class="text-icon" />
+							<span class="text-subtle pointer-events-none absolute top-0 right-1.5 py-2.5 text-xs">
+								<BiChevronDown aria-hidden="true" />
 							</span>
 						</div>
 					</div>
 				{/if}
 			</div>
-			<ol class="flex flex-col gap-0.5 md:px-0">
+			<ol class="flex flex-col gap-0.5 lg:px-0">
 				{#each searchResult.items as item (item['@id'])}
 					<li>
 						<SearchCard {item} />
@@ -205,6 +217,8 @@
 {/if}
 
 <style lang="postcss">
+	@reference "../../../app.css";
+
 	.toolbar {
 		@apply grid;
 		grid-template-areas:
@@ -227,6 +241,10 @@
 		grid-template-areas: 'filters results';
 	}
 
+	.filter-modal-toggle {
+		grid-area: filter-modal-toggle;
+	}
+
 	.filters {
 		grid-area: filters;
 	}
@@ -244,9 +262,7 @@
 		grid-area: sort-select;
 
 		& select {
-			@apply appearance-none pr-6 text-right;
-			/* Safari text-align fix */
-			text-align-last: right;
+			@apply appearance-none px-8;
 		}
 	}
 
@@ -258,7 +274,13 @@
 		grid-area: search-related;
 	}
 
-	@media screen and (min-width: theme('screens.sm')) {
+	.tab-selected {
+		padding-bottom: cacl(--var(--spacing) * 3.5);
+		border-bottom-color: var(--color-accent);
+		border-bottom-width: 0.125rem;
+	}
+
+	@variant sm {
 		.toolbar {
 			grid-template-areas:
 				'filter-modal-toggle search-related'
@@ -267,13 +289,9 @@
 		}
 	}
 
-	@media screen and (min-width: theme('screens.md')) {
+	@variant lg {
 		.filters {
 			display: block;
-		}
-
-		.filter-modal-toggle {
-			display: none;
 		}
 
 		.toolbar {
@@ -281,18 +299,5 @@
 				'search-related search-related'
 				'hits sort-select';
 		}
-	}
-	.tab-header {
-		@apply block py-4;
-	}
-
-	.tab {
-		@apply block py-4 pl-4 pr-3.5 lowercase no-underline;
-		transition: filter 0.1s ease;
-	}
-
-	.tab-selected {
-		@apply border-primary pb-3.5;
-		border-bottom-width: 0.125rem;
 	}
 </style>
