@@ -1,8 +1,8 @@
 <script lang="ts">
-	import { page } from '$app/stores';
-	import { afterNavigate } from '$app/navigation';
-	import { SuperSearch, lxlQualifierPlugin } from 'supersearch';
-	import SuggestionCard from './SuggestionCard.svelte';
+	import { page } from '$app/state';
+	import { afterNavigate, goto } from '$app/navigation';
+	import { SuperSearch, lxlQualifierPlugin, type Selection } from 'supersearch';
+	import Suggestion from './Suggestion.svelte';
 	import addDefaultSearchParams from '$lib/utils/addDefaultSearchParams';
 	import getSortedSearchParams from '$lib/utils/getSortedSearchParams';
 	import getLabelFromMappings from '$lib/utils/getLabelsFromMapping.svelte';
@@ -11,7 +11,7 @@
 	import { lxlQuery } from 'codemirror-lang-lxlquery';
 	import BiXLg from '~icons/bi/x-lg';
 	import BiArrowLeft from '~icons/bi/arrow-left';
-	import BiArrowUpLeft from '~icons/bi/arrow-up-left';
+	import BiSearch from '~icons/bi/search';
 	import BiChevronDown from '~icons/bi/chevron-down';
 	import BiChevronUp from '~icons/bi/chevron-up';
 	import '$lib/styles/lxlquery.css';
@@ -22,30 +22,32 @@
 
 	let { placeholder = '' }: Props = $props();
 	let q = $state(
-		$page.params.fnurgel
+		page.params.fnurgel
 			? ''
-			: addSpaceIfEndingQualifier($page.url.searchParams.get('_q')?.trim() || '')
+			: addSpaceIfEndingQualifier(page.url.searchParams.get('_q')?.trim() || '')
 	);
-	let cursor: number = $state(0);
+	let selection: Selection | undefined = $state();
+	let cursor = $derived(selection?.head || 0);
+
 	let superSearch = $state<ReturnType<typeof SuperSearch>>();
 	let showMoreFilters = $state(false);
 
-	let params = getSortedSearchParams(addDefaultSearchParams($page.url.searchParams));
-	// Always reset these params on new search
-	params.set('_offset', '0');
-	params.delete('_i');
-	params.delete('_o');
-	params.delete('_p');
-	const searchParams = Array.from(params);
-
+	let pageParams = $derived.by(() => {
+		let p = getSortedSearchParams(addDefaultSearchParams(page.url.searchParams));
+		// Always reset these params on new search
+		p.set('_offset', '0');
+		p.delete('_i');
+		p.delete('_o');
+		p.delete('_p');
+		return p;
+	});
 	let suggestMapping: DisplayMapping[] | undefined = $state();
 
 	afterNavigate(({ to }) => {
 		/** Update input value after navigation on /find route */
 		if (to?.url) {
-			q = $page.params.fnurgel
-				? ''
-				: addSpaceIfEndingQualifier(new URL(to.url).searchParams.get('_q')?.trim() || '');
+			const toQ = addSpaceIfEndingQualifier(new URL(to.url).searchParams.get('_q')?.trim() || '');
+			q = page.params.fnurgel ? '' : toQ !== '*' ? toQ : ''; // hide wildcard in input field
 			superSearch?.hideExpandedSearch();
 		}
 	});
@@ -70,12 +72,19 @@
 		return undefined;
 	}
 
-	function handleShouldShowStartContent(value: string, cursor: number) {
-		const tree = lxlQuery.language.parser.parse(value);
-		const node = tree.resolveInner(cursor, -1);
+	function handleShouldShowStartContent(value: string, selection?: Selection) {
+		if (selection && selection.anchor == selection.head) {
+			const tree = lxlQuery.language.parser.parse(value);
+			const nodeLeft = tree.resolveInner(selection.head, -1);
+			const nodeRight = tree.resolveInner(selection.head, 1);
 
-		/** Start content should be shown when the cursor isn't placed inside a qualifier or edited string part */
-		if (!node.parent?.name) {
+			/** Start content should be shown when the cursor isn't placed inside a qualifier or edited string part */
+			if (!nodeLeft.parent?.name && !nodeRight.parent?.name) {
+				return true;
+			}
+		}
+
+		if (!value) {
 			return true;
 		}
 
@@ -88,19 +97,19 @@
 	}
 
 	function addQualifierKey(qualifierKey: string) {
+		superSearch?.showExpandedSearch(); // keep dialog open (since 'regular' search is hidden on mobile)
 		superSearch?.dispatchChange({
 			change: {
 				from: cursor,
 				to: cursor,
-				insert: `${qualifierKey}:""`
+				insert: `${qualifierKey}:`
 			},
 			selection: {
-				anchor: cursor + qualifierKey?.length + 2,
-				head: cursor + qualifierKey?.length + 2
+				anchor: cursor + qualifierKey?.length + 1,
+				head: cursor + qualifierKey?.length + 1
 			},
 			userEvent: 'input.complete'
 		});
-		superSearch?.hideExpandedSearch();
 	}
 
 	function addQualifier(qualifier: QualifierSuggestion) {
@@ -109,24 +118,41 @@
 			selection: { anchor: qualifier.cursor + 1, head: qualifier.cursor + 1 },
 			userEvent: 'input.complete'
 		});
+		goto(getFullQualifierLink(qualifier._q));
+	}
+
+	function removeQualifier(qualifier: string) {
+		const newQ = addSpaceIfEndingQualifier(q.replace(qualifier, '').trim());
+		const insertCursor = Math.min(q.indexOf(qualifier), newQ.length);
+		const newSearchParams = new URLSearchParams(pageParams);
+		newSearchParams.set('_q', newQ.trim() ? newQ : '*');
+
+		superSearch?.dispatchChange({
+			change: { from: 0, to: q.length, insert: newQ },
+			selection: { anchor: insertCursor, head: insertCursor },
+			userEvent: 'delete'
+		});
+		goto('/find?' + newSearchParams.toString());
 	}
 
 	let derivedLxlQualifierPlugin = $derived.by(() => {
 		function getLabels(key: string, value?: string) {
-			let pageMapping = $page.data.searchResult?.mapping;
+			let pageMapping = page.data.searchResult?.mapping;
 			return getLabelFromMappings(key, value, pageMapping, suggestMapping);
 		}
-		return lxlQualifierPlugin(getLabels);
+		return lxlQualifierPlugin(getLabels, removeQualifier);
 	});
 
 	let moreFiltersRowIndex = $derived(showMoreFilters ? 7 : 4);
 
 	function getFullQualifierLink(q: string) {
-		const params = new URLSearchParams($page.url.searchParams.toString());
-		params.set('_q', q);
-		params.delete('_i');
-		params.set('_offset', '0');
-		return `/find?${params.toString()}`;
+		const newParams = new URLSearchParams(pageParams);
+		newParams.set('_q', q);
+		return `/find?${newParams.toString()}`;
+	}
+
+	export function showExpandedSearch() {
+		superSearch?.showExpandedSearch();
 	}
 </script>
 
@@ -147,33 +173,32 @@
 	isFocusedRow: (rowIndex: number) => boolean;
 	rowIndex: number;
 })}
-	<div role="row" class="start-item" class:focused={isFocusedRow(rowIndex)}>
+	<div role="row" class="suggestion" class:focused={isFocusedRow(rowIndex)}>
 		<button
 			type="button"
 			role="gridcell"
 			id={getCellId(rowIndex, 0)}
-			class="flex min-h-12 w-full items-center px-4 hover:bg-main"
+			class="hover:bg-primary-50 flex min-h-12 w-full items-center px-4"
 			class:focused-cell={isFocusedCell(rowIndex, 0)}
 			onclick={() => addQualifierKey(qualifierKey)}
 		>
-			<span class="overflow-hidden text-ellipsis whitespace-nowrap">
-				<strong class="text-3-cond-bold">{qualifierLabel}:</strong>
-				<span class="text-sm italic text-tertiary">{qualifierPlaceholder}</span>
+			<span class="truncate text-sm">
+				<span class="font-medium">{qualifierLabel}:</span>
+				<span class="text-subtle italic">{qualifierPlaceholder}</span>
 			</span>
-			<BiArrowUpLeft class="ml-auto inline-block" />
 		</button>
 	</div>
 {/snippet}
 
-<form class="relative w-full" action="find" onsubmit={handleSubmit}>
+<form class="relative w-full" action="find" onsubmit={handleSubmit} data-testid="main-search">
 	<SuperSearch
 		name="_q"
 		bind:this={superSearch}
 		bind:value={q}
-		bind:cursor
+		bind:selection
 		language={lxlQuery}
 		{placeholder}
-		endpoint={`/api/${$page.data.locale}/supersearch`}
+		endpoint={`/api/${page.data.locale}/supersearch`}
 		queryFn={(query, cursor) => {
 			return new URLSearchParams({
 				_q: query,
@@ -186,38 +211,44 @@
 		shouldShowStartContentFn={handleShouldShowStartContent}
 		extensions={[derivedLxlQualifierPlugin]}
 		toggleWithKeyboardShortcut
-		comboboxAriaLabel={$page.data.t('search.search')}
-		defaultResultRow={1}
+		comboboxAriaLabel={page.data.t('search.search')}
 		defaultInputCol={2}
+		loadMoreLabel={page.data.t('search.showMore')}
+		debouncedWait={100}
 	>
 		{#snippet loadingIndicator()}
-			<div class="flex min-h-11 w-full items-center px-4 text-left">
+			<!-- <div class="flex min-h-11 w-full items-center px-4 text-left">
 				{$page.data.t('search.loading')}
-			</div>
+			</div> -->
 		{/snippet}
 		{#snippet inputRow({
 			expanded,
 			inputField,
 			getCellId,
 			isFocusedCell,
-			onclickSubmit,
+			// onclickSubmit,
 			onclickClear,
 			onclickClose
 		})}
-			<div class="supersearch-input">
+			<div
+				class="supersearch-input rounded-d bg-input outline-primary-200 has-focus:outline-primary-600 flex min-h-12 w-full cursor-text overflow-hidden rounded-md outline focus-within:relative"
+			>
 				{#if expanded}
 					<button
 						type="button"
 						id={getCellId(0)}
 						class:focused-cell={isFocusedCell(0)}
-						aria-label={$page.data.t('general.close')}
-						class="button-ghost min-h-12 min-w-11 rounded-none border-none hover:bg-main sm:hidden"
+						aria-label={page.data.t('general.close')}
+						class="p-4 sm:hidden"
 						onclick={onclickClose}
 					>
 						<BiArrowLeft />
 					</button>
 				{/if}
 				<div class="flex-1 overflow-hidden">
+					<div class={['text-subtle absolute p-4', expanded ? 'hidden sm:block' : 'block']}>
+						<BiSearch aria-hidden="true" />
+					</div>
 					{@render inputField()}
 				</div>
 				{#if q}
@@ -225,52 +256,52 @@
 						type="reset"
 						id={getCellId(1)}
 						class:focused-cell={isFocusedCell(1)}
-						class="button-ghost min-h-12 rounded-none border-none hover:bg-main"
-						aria-label={$page.data.t('search.clearFilters')}
+						class="text-subtle p-4"
+						aria-label={page.data.t('search.clearFilters')}
 						onclick={onclickClear}
 					>
 						<BiXLg />
 					</button>
 				{/if}
-				<button
+				<!-- <button
 					type="submit"
 					id={getCellId(2)}
 					class:focused-cell={isFocusedCell(2)}
-					class="submit-action button-primary min-h-12 rounded-none"
+					class="submit-action min-h-12 rounded-none"
 					enterkeyhint="search"
 					onclick={onclickSubmit}
 				>
 					{$page.data.t('search.search')}
-				</button>
+				</button> -->
 			</div>
 		{/snippet}
 		{#snippet startContent({ getCellId, isFocusedCell, isFocusedRow })}
 			<div role="rowgroup">
-				<div class="flex min-h-11 w-full items-center px-4 text-3-cond-bold">
-					{$page.data.t('search.supersearchStartHeader')}
+				<div class="text-2xs text-subtle flex w-full items-center px-4 py-2 font-medium">
+					{page.data.t('search.supersearchStartHeader')}
 				</div>
 				{@render startFilterItem({
-					qualifierKey: $page.data.t('qualifiers.contributorKey'),
-					qualifierLabel: $page.data.t('qualifiers.contributorLabel'),
-					qualifierPlaceholder: $page.data.t('qualifiers.contributorPlaceholder'),
+					qualifierKey: page.data.t('qualifiers.contributorKey'),
+					qualifierLabel: page.data.t('qualifiers.contributorLabel'),
+					qualifierPlaceholder: page.data.t('qualifiers.contributorPlaceholder'),
 					getCellId,
 					isFocusedCell,
 					isFocusedRow,
 					rowIndex: 1
 				})}
 				{@render startFilterItem({
-					qualifierKey: $page.data.t('qualifiers.titleKey'),
-					qualifierLabel: $page.data.t('qualifiers.titleLabel'),
-					qualifierPlaceholder: $page.data.t('qualifiers.titlePlaceholder'),
+					qualifierKey: page.data.t('qualifiers.titleKey'),
+					qualifierLabel: page.data.t('qualifiers.titleLabel'),
+					qualifierPlaceholder: page.data.t('qualifiers.titlePlaceholder'),
 					getCellId,
 					isFocusedCell,
 					isFocusedRow,
 					rowIndex: 2
 				})}
 				{@render startFilterItem({
-					qualifierKey: $page.data.t('qualifiers.languageKey'),
-					qualifierLabel: $page.data.t('qualifiers.languageLabel'),
-					qualifierPlaceholder: $page.data.t('qualifiers.languagePlaceholder'),
+					qualifierKey: page.data.t('qualifiers.languageKey'),
+					qualifierLabel: page.data.t('qualifiers.languageLabel'),
+					qualifierPlaceholder: page.data.t('qualifiers.languagePlaceholder'),
 					getCellId,
 					isFocusedCell,
 					isFocusedRow,
@@ -278,27 +309,27 @@
 				})}
 				{#if showMoreFilters}
 					{@render startFilterItem({
-						qualifierKey: $page.data.t('qualifiers.subjectKey'),
-						qualifierLabel: $page.data.t('qualifiers.subjectLabel'),
-						qualifierPlaceholder: $page.data.t('qualifiers.subjectPlaceholder'),
+						qualifierKey: page.data.t('qualifiers.subjectKey'),
+						qualifierLabel: page.data.t('qualifiers.subjectLabel'),
+						qualifierPlaceholder: page.data.t('qualifiers.subjectPlaceholder'),
 						getCellId,
 						isFocusedCell,
 						isFocusedRow,
 						rowIndex: 4
 					})}
 					{@render startFilterItem({
-						qualifierKey: $page.data.t('qualifiers.yearKey'),
-						qualifierLabel: $page.data.t('qualifiers.yearLabel'),
-						qualifierPlaceholder: $page.data.t('qualifiers.yearPlaceholder'),
+						qualifierKey: page.data.t('qualifiers.yearKey'),
+						qualifierLabel: page.data.t('qualifiers.yearLabel'),
+						qualifierPlaceholder: page.data.t('qualifiers.yearPlaceholder'),
 						getCellId,
 						isFocusedCell,
 						isFocusedRow,
 						rowIndex: 5
 					})}
 					{@render startFilterItem({
-						qualifierKey: $page.data.t('qualifiers.genreFormKey'),
-						qualifierLabel: $page.data.t('qualifiers.genreFormLabel'),
-						qualifierPlaceholder: $page.data.t('qualifiers.genreFormPlaceholder'),
+						qualifierKey: page.data.t('qualifiers.genreFormKey'),
+						qualifierLabel: page.data.t('qualifiers.genreFormLabel'),
+						qualifierPlaceholder: page.data.t('qualifiers.genreFormPlaceholder'),
 						getCellId,
 						isFocusedCell,
 						isFocusedRow,
@@ -310,69 +341,28 @@
 						type="button"
 						role="gridcell"
 						id={getCellId(moreFiltersRowIndex, 0)}
-						class="flex min-h-11 w-full items-center px-4 hover:bg-main"
+						class="hover:bg-primary-50 flex min-h-11 w-full items-center px-4 text-sm"
 						class:focused-cell={isFocusedCell(moreFiltersRowIndex, 0)}
 						onclick={() => (showMoreFilters = !showMoreFilters)}
 					>
 						{#if showMoreFilters}
-							<BiChevronUp class="mr-2" />
-							{$page.data.t('search.showFewer')}
+							<BiChevronUp class="text-subtle mr-2" />
+							{page.data.t('search.showFewer')}
 						{:else}
-							<BiChevronDown class="mr-2" />
-							{$page.data.t('search.showMore')}
+							<BiChevronDown class="text-subtle mr-2" />
+							{page.data.t('search.showMore')}
 						{/if}
 					</button>
 				</div>
 			</div>
 		{/snippet}
-		{#snippet persistentResultItemRow({ getCellId, isFocusedCell })}
-			<div class="flex min-h-12 w-full items-stretch border-b border-b-primary/16">
-				<button
-					type="submit"
-					role="gridcell"
-					id={getCellId(0)}
-					class="flex flex-1 items-center px-4 no-underline text-3-cond-bold hover:bg-main"
-					class:focused-cell={isFocusedCell(0)}
-					enterkeyhint="search"
-					>{$page.data.t('search.showAllResults')}
-				</button>
-			</div>
-		{/snippet}
 		{#snippet resultItemRow({ resultItem, getCellId, isFocusedCell })}
 			{#if resultItem}
-				<div class="suggestion-card-container flex text-xs md:text-sm">
-					<SuggestionCard item={resultItem} cellId={getCellId(0)} isFocused={isFocusedCell(0)} />
-					{#if resultItem.qualifiers.length}
-						<div class="flex pr-2">
-							{#each resultItem.qualifiers as qualifier, index}
-								<!-- pills -->
-								<a
-									id={getCellId(index + 1)}
-									role="gridcell"
-									class={[
-										'flex items-center p-1 no-underline hover:bg-main sm:p-1.5',
-										isFocusedCell(index + 1) && 'focused-cell'
-									]}
-									onclick={() => addQualifier(qualifier)}
-									href={getFullQualifierLink(qualifier._q)}
-								>
-									<span class="lxl-qualifier atomic add-qualifier">
-										<BiArrowUpLeft
-											class=""
-											fill="currentColor"
-											aria-label={$page.data.t('search.addAs')}
-										/>
-										<span class="whitespace-nowrap first-letter:capitalize">{qualifier.label}</span>
-									</span>
-								</a>
-							{/each}
-						</div>
-					{/if}
-				</div>
+				<Suggestion item={resultItem} {getCellId} {isFocusedCell} {addQualifier} />
 			{/if}
 		{/snippet}
 	</SuperSearch>
-	{#each searchParams as [name, value]}
+	{#each Array.from(pageParams) as [name, value] (name)}
 		{#if name !== '_q'}
 			<input type="hidden" {name} {value} />
 		{/if}
@@ -380,71 +370,78 @@
 </form>
 
 <style lang="postcss">
-	.supersearch-input {
-		@apply relative flex min-h-12 w-full cursor-text overflow-hidden rounded-md bg-cards focus-within:outline focus-within:outline-2 focus-within:outline-accent-dark/32;
-	}
+	@reference "../../../app.css";
 
-	.suggestion-card-container {
-		container-type: inline-size;
-	}
-
-	.add-qualifier {
-		@apply flex items-center gap-1 rounded-sm px-2 py-1;
+	/* search */
+	:global(#supersearch) {
+		display: none;
+		@variant sm {
+			display: block;
+		}
 	}
 
 	/* dialog */
 
 	:global(.supersearch-dialog) {
-		@apply static m-0 h-full max-h-screen w-full max-w-full border-none bg-transparent p-0;
+		@apply static m-0 h-full max-h-screen w-full max-w-full bg-transparent p-0;
+		top: var(--offset-top, 0);
 	}
 
 	:global(.supersearch-dialog-wrapper) {
-		@apply pointer-events-none header-layout;
+		@apply header-layout pointer-events-none px-0 sm:px-6 lg:px-2;
 		grid-template-areas: 'supersearch-content supersearch-content supersearch-content';
 
-		@media screen and (min-width: theme('screens.sm')) {
+		@variant sm {
 			grid-template-areas: '. supersearch-content .';
 		}
 	}
 
 	:global(.supersearch-dialog-content) {
-		@apply pointer-events-auto max-h-screen overflow-hidden overflow-y-scroll rounded-md bg-cards drop-shadow-md;
+		@apply pointer-events-auto max-h-screen overflow-hidden overflow-y-scroll rounded-md drop-shadow-md;
+		background: var(--color-page);
 		grid-area: supersearch-content;
 		scrollbar-width: none;
 	}
 
 	:global(.supersearch-dialog .supersearch-combobox) {
-		@apply sticky top-0 z-20 items-stretch border-b border-b-primary/16 bg-cards p-4;
+		@apply sticky top-0 z-20 items-stretch px-4 pt-3 pb-2;
+		background-color: var(--color-page);
+	}
+
+	:global(.supersearch-suggestions) {
+		@apply min-h-2;
 	}
 
 	:global(.supersearch-dialog .supersearch-input) {
 		@apply overflow-hidden rounded-md sm:px-0;
-		box-shadow: inset 0 0 0 1px rgba(105, 65, 25, 0.24);
 	}
 
 	:global(.supersearch-dialog .focused) {
-		@apply bg-site-header/24;
+		background-color: var(--color-primary-100);
 	}
 
 	:global(.focused-cell) {
-		@apply bg-site-header/40;
+		background-color: var(--color-primary-100);
 	}
 
-	:global(.button-primary.focused-cell) {
+	/* :global(.button-primary.focused-cell) {
 		@apply before:opacity-100;
-	}
+	} */
 
 	/* suggestions */
 
 	:global(.supersearch-suggestions [role='row']:last-child) {
-		/* border-bottom: 1px solid rgb(var(--color-primary) / 0.12); */
-		@apply border-b border-b-primary/16;
+		border-color: var(--color-neutral);
 	}
 
 	/* snippets elements */
 
 	:global(.supersearch-show-more) {
-		@apply flex min-h-11 w-full items-center px-4 text-left hover:bg-main;
+		@apply flex min-h-11 w-full items-center px-4 text-left text-sm;
+
+		&:hover {
+			background-color: var(--color-primary-50);
+		}
 	}
 
 	/* codemirror elements */
@@ -454,14 +451,22 @@
 	}
 
 	:global(.codemirror-container .cm-scroller) {
-		@apply min-h-12 font-sans outline-none text-3-regular;
+		@apply min-h-12 font-sans outline-hidden;
 		scrollbar-width: none;
 	}
 
 	:global(.supersearch-input .cm-line) {
-		@apply px-4;
+		padding-left: calc(var(--spacing, 0.25rem) * 12);
 		min-height: 28px;
 		line-height: 28px;
+		font-size: var(--text-xs);
+	}
+
+	:global(.supersearch-dialog .supersearch-input .cm-line) {
+		padding-left: 0;
+		@variant sm {
+			padding-left: calc(var(--spacing, 0.25rem) * 12);
+		}
 	}
 
 	:global(.supersearch-input .cm-focused) {
@@ -472,5 +477,9 @@
 		padding-top: 0.6125rem;
 		padding-bottom: 0.6125rem;
 		outline: none;
+	}
+
+	:global(.codemirror-container .cm-placeholder) {
+		color: var(--color-placeholder);
 	}
 </style>
