@@ -18,15 +18,13 @@ function getLabelFromMappings(
 	const keyLabel = suggestLabels.keyLabel || pageLabels.keyLabel;
 	const valueLabel = suggestLabels.valueLabel || pageLabels.valueLabel;
 	const invalid = suggestLabels.invalid || pageLabels.invalid;
-	// only page data have 'up' links we can use
-	const removeLink = pageLabels.keyLabel ? pageLabels.removeLink : undefined;
 
 	if (suggestMapping?.length) {
 		// save latest mapping as fallback for error responses etc
 		prevSuggestMapping = suggestMapping;
 	}
 
-	return { keyLabel, valueLabel, removeLink, invalid };
+	return { keyLabel, valueLabel, invalid };
 }
 
 function iterateMapping(
@@ -36,7 +34,6 @@ function iterateMapping(
 ) {
 	let keyLabel: string | undefined;
 	let valueLabel: string | undefined;
-	let removeLink: string | undefined;
 	let invalid: boolean = false;
 
 	if (mapping && Array.isArray(mapping)) {
@@ -56,14 +53,15 @@ function iterateMapping(
 					if (isLinked && value === el?._value && el?.displayStr) {
 						// only use atomic ranges for linked values
 						valueLabel = el.displayStr;
-						// only show remove btn for pills that can't be edited
-						removeLink = el.up?.['@id'];
 					}
+				} else if (!key && value === el?._value && el?.displayStr) {
+					// ...unless a filter alias (no key, only value)
+					valueLabel = el.displayStr;
 				}
 			});
 		}
 	}
-	return { keyLabel, valueLabel, removeLink, invalid };
+	return { keyLabel, valueLabel, invalid };
 }
 
 export default getLabelFromMappings;
