@@ -4,12 +4,10 @@
 	import * as CookieConsent from 'vanilla-cookieconsent';
 </script>
 
-<footer
-	class="mt-auto flex flex-col justify-between gap-8 bg-primary/4 p-8 text-secondary text-3-regular sm:flex-row lg:p-16"
->
-	<div class="flex flex-col gap-8 sm:flex-row sm:gap-16">
-		<nav class="flex flex-col gap-4" aria-labelledby="nav-info">
-			<p id="nav-info" class="text-primary text-4-cond-bold">
+<footer class="mt-auto flex flex-col justify-between gap-8 bg-neutral-100 p-8 sm:flex-row 2xl:p-10">
+	<div class="flex flex-col gap-4 sm:flex-row sm:gap-16 [&_li>*]:text-sm [&_p]:font-medium">
+		<nav class="flex flex-col gap-2" aria-labelledby="nav-info">
+			<p id="nav-info">
 				{$page.data.t('footer.information')}
 			</p>
 			<ul>
@@ -20,14 +18,12 @@
 					>
 				</li>
 				<li>
-					<button class="underline" on:click={CookieConsent.showPreferences}
-						>{$page.data.t('footer.cookies')}</button
-					>
+					<button on:click={CookieConsent.showPreferences}>{$page.data.t('footer.cookies')}</button>
 				</li>
 			</ul>
 		</nav>
-		<nav class="flex flex-col gap-4" aria-labelledby="nav-contact">
-			<p id="nav-contact" class="text-primary text-4-cond-bold">{$page.data.t('footer.contact')}</p>
+		<nav class="flex flex-col gap-2" aria-labelledby="nav-contact">
+			<p id="nav-contact">{$page.data.t('footer.contact')}</p>
 			<ul>
 				<li>
 					<a href="https://www.kb.se/om-oss/kontakta-oss/#item-47bef1f3916eca87a61b1eed"
@@ -41,6 +37,6 @@
 		</nav>
 	</div>
 	<div class="flex items-end">
-		<img class="h-20 w-auto md:h-24 lg:h-28" alt={$page.data.t('footer.logo')} src={KbLogo} />
+		<img class="h-20 w-auto lg:h-20 2xl:h-24" alt={$page.data.t('footer.logo')} src={KbLogo} />
 	</div>
 </footer>
