@@ -8,4 +8,7 @@
 	<title>{getPageTitle($page.url.searchParams.get('_q')?.trim())}</title>
 </svelte:head>
 <h1 class="sr-only">{$page.data.t('search.searchResults')}</h1>
-<SearchResult searchResult={$page.data.searchResult} />
+<SearchResult
+	searchResult={$page.data.searchResult}
+	showMapping={$page.url.searchParams.get('_x') === 'mappings'}
+/>
